Added logic to retrieve primary group for users
In Active Directory, the primary group is not stored as other groups are per
user. Instead, AD relies on the PrimaryGroupID attribute on a user object to
identity the primary group. However, in order to be able to retrieve the group
object from the server, the SID of the user must be used to compute the group's
SID.

Foregoing a more technical explanation as to the semantics and meaning behind
the Active Directory implementation, what follows is a brief overview of the
added functionality.

First, the SID of the user must be converted from a binary format to a string
format.

The resultant string representation of the SID looks somewhat like this:

```
S-1-5-20-123456789-1234567890-123456789-1234
```

The last section of the SID, in this case, `1234`, identifies the particular
user. This portion of the SID is removed and replaced with the PrimaryGroupID.

Finally, the new SID is queried against AD to retrieve the user's primary group.